Blacklist checker

+++ src/main/resources/static/js/customerOrNotScript.js
@@ -2,47 +2,63 @@ function submitBooking() {
     const name = document.getElementById("name").value;
     const email = document.getElementById("email").value;
 
-    localStorage.setItem("name", name);
-    localStorage.setItem("email", email);
+    let xhr1 = new XMLHttpRequest();
+    let url1 = "/customer/blacklisted/" + email;
 
-    //Fan måste ju skicka inställningarna också
-    let startDate = localStorage.getItem("startDate");
-    let endDate = localStorage.getItem("endDate");
+    xhr1.open("GET", url1, true);
+    xhr1.send();
 
-    //Hämta chosenRooms från localstorage
-    let chosenRooms = JSON.parse(localStorage.getItem("chosenRooms")) || [];
+    xhr1.onreadystatechange = function () {
+        if (xhr1.readyState === 4 && xhr1.status === 200) {
+            if (!xhr1.responseText.includes("bookingConfirmation")) {
+                const parser = new DOMParser();
+                const htmlDoc = parser.parseFromString(xhr1.responseText, 'text/html');
+                document.open();
+                document.write(htmlDoc.documentElement.outerHTML);
+                document.close();
+            } else {
+                localStorage.setItem("name", name);
+                localStorage.setItem("email", email);
 
-    //Vet inte om detta är så bra idk
-    let bookingData = {
-        id: -1,
-        name: name,
-        email: email,
-        startDate: startDate,
-        endDate: endDate,
-        chosenRooms: chosenRooms
-    };
+                //Fan måste ju skicka inställningarna också
+                let startDate = localStorage.getItem("startDate");
+                let endDate = localStorage.getItem("endDate");
 
-    //Stoppa i body i anrop  @PostMapping("/submitBooking")
-    let xhr = new XMLHttpRequest();
-    let url = "/submitBookingCustomer";
+                //Hämta chosenRooms från localstorage
+                let chosenRooms = JSON.parse(localStorage.getItem("chosenRooms")) || [];
 
-    xhr.open("POST", url, true);
-    xhr.setRequestHeader("Content-Type", "application/json");
+                //Vet inte om detta är så bra idk
+                let bookingData = {
+                    id: -1,
+                    name: name,
+                    email: email,
+                    startDate: startDate,
+                    endDate: endDate,
+                    chosenRooms: chosenRooms
+                };
 
-    let data = JSON.stringify(bookingData);
-    console.log("data = " + data);
-    xhr.send(data);
+                //Stoppa i body i anrop  @PostMapping("/submitBooking")
+                let xhr2 = new XMLHttpRequest();
+                let url2 = "/submitBookingCustomer";
 
-    xhr.onreadystatechange = function () {
-        if (xhr.readyState === 4 && xhr.status === 200) {
-            console.log(xhr.responseText);
-            const parser = new DOMParser();
-            const htmlDoc = parser.parseFromString(xhr.responseText, 'text/html');
-            document.open();
-            document.write(htmlDoc.documentElement.outerHTML);
-            document.close();
+                xhr2.open("POST", url2, true);
+                xhr2.setRequestHeader("Content-Type", "application/json");
+
+                let data = JSON.stringify(bookingData);
+                console.log("data = " + data);
+                xhr2.send(data);
+
+                xhr2.onreadystatechange = function () {
+                    if (xhr2.readyState === 4 && xhr2.status === 200) {
+                        console.log(xhr2.responseText);
+                        const parser = new DOMParser();
+                        const htmlDoc = parser.parseFromString(xhr2.responseText, 'text/html');
+                        document.open();
+                        document.write(htmlDoc.documentElement.outerHTML);
+                        document.close();
+                    }
+                }
+            }
         }
     }
 }
-
-
